remove person test base class
it's a potentially a bit confusing and doesn't serve much purpose other
than excessive de-duplication of code
